+++ src/test/java/demo.java
@@ -15,10 +15,10 @@ public class demo {
 		
 		if (! root.isParent()) return;
 		
-		for (Node contact : root.getNodes().get("contact").get(n -> n.isParent())) {
+		for (Node contact : root.getNodes().find("contact").find(n -> n.isParent())) {
 			
-			Node name = contact.getNodes().get("firstname").get(1);
-			NodeSet numbers = contact.getNodes().get("phonenumber");
+			Node name = contact.getNodes().get("firstname");
+			NodeSet numbers = contact.getNodes().find("phonenumber");
 			
 			System.out.println(name.getValue() + " has " + numbers.size() + " phone number(s).");